Update BaanChang.lua
Doubled production and culture from Units gifted by CS.  5 per era scales too poorly, and is a very slight bonus.

I think the bonus should be given to the city with the baan chang though, not all given to the capital alone. The timing of CS gifts is easy for players to guess, and could be abused to rush wonders, etc.